requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.128
botocore==1.34.128
cachetools==5.3.3
cellpose==3.0.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dominate==2.9.1
fastremap==1.14.1
filelock==3.13.3
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.4.0
gdown==5.2.0
google==3.0.0
google-api-core==2.18.0
google-api-python-client==2.134.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleAPI==0.1.0
googleapis-common-protos==1.63.0
grpcio==1.64.1
gunicorn==22.0.0
h5py==3.11.0
httplib2==0.22.0
idna==3.6
imagecodecs==2024.1.1
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
natsort==8.4.0
networkx==3.2.1
numba==0.59.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pyparsing==3.1.2
PyQt6==6.6.1
PyQt6-Qt6==6.6.2
PyQt6-sip==13.6.0
pyqtgraph==0.13.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
QtPy==2.4.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
roifile==2024.3.20
rsa==4.9
s3transfer==0.10.1
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
soupsieve==2.5
superqt==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.15.0
termcolor==2.4.0
tifffile==2024.2.12
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.19
visitor==0.1.3
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.18.1
